+++ backend-server/application/src/main/java/com/apitable/shared/util/page/PageInfo.java
@@ -33,23 +33,23 @@ import lombok.Data;
 @Data
 public class PageInfo<T> {
 
-    private int pageNum;
+    private long pageNum;
 
-    private int pageSize;
+    private long pageSize;
 
-    private int size;
+    private long size;
 
-    private int total;
+    private long total;
 
-    private int pages;
+    private long pages;
 
-    private int startRow;
+    private long startRow;
 
-    private int endRow;
+    private long endRow;
 
-    private int prePage;
+    private long prePage;
 
-    private int nextPage;
+    private long nextPage;
 
     private Boolean firstPage = false;
 
@@ -73,7 +73,7 @@ public class PageInfo<T> {
      * @param total    total records
      * @param records  records
      */
-    public PageInfo(int pageNum, int pageSize, int total, List<T> records) {
+    public PageInfo(long pageNum, long pageSize, long total, List<T> records) {
         this.pageNum = pageNum;
         this.pageSize = pageSize;
         this.total = total;
@@ -91,7 +91,7 @@ public class PageInfo<T> {
             this.size = 0;
             return;
         }
-        int sub = this.total - ((this.pageNum - 1) * this.pageSize);
+        long sub = this.total - ((this.pageNum - 1) * this.pageSize);
         if (sub <= 0) {
             this.size = 0;
             return;
@@ -101,7 +101,7 @@ public class PageInfo<T> {
             this.size = records.size();
             return;
         }
-        this.records = records.subList(0, sub);
+        this.records = records.subList(0, (int) sub);
         this.size = sub;
     }
 
@@ -109,7 +109,7 @@ public class PageInfo<T> {
         if (getPageSize() == 0) {
             this.pages = 0;
         }
-        int pages = getTotal() / getPageSize();
+        long pages = getTotal() / getPageSize();
         if (getTotal() % getPageSize() != 0) {
             pages++;
         }
